vir/vir/mental-health-app/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow==2.19.0
numpy
opencv-python==4.8.1.78
librosa==0.10.1
scikit-learn==1.3.2
PyJWT==2.8.0
bcrypt==4.1.2
pillow==10.1.0